a7xelte: Add support for Samsung's LPM binary
'lpm' is Samsung's equivalent for the AOSP's 'charger'. This handles the
screen interface when charging the phone while it's powered off.

Aside from the proprietary nature of this, there are a couple of features
that is not present on AOSP's/LOS' charger binary:
1. Fast charging indicator
2. Battery percentages while charging.
3. Colored animations (AOSP's/LOS' 'charger', by defaultm is monochrome)
4. You can 'wake' the screen by pressing the Home button.
5. You can trigger a reboot by holding the power button.

To integrate this, we need to add rules to allow lpm in enforcing mode.
That should be addressed in the previous patch.

Next the lpm binary and all of the files it requires must be present.
Because we will place 'lpm' in /vendor/ rather than /system/, it has to be
edited to redirect the /system/ paths to /vendor/.

Lastly, we need to write an init script to run the lpm binary when the
phone enters 'charger' mode. Coincidentally, because 'charger' exists
in the same system, we need to stop it before running 'lpm' in its stead.
In Samsung's default ROMs, 'charger' is not built.

Signed-off-by: Mark Enriquez <bugmenot.oss@gmail.com>
Change-Id: Idf307c931c3164add08a97f08ac4117e8ed9f78b