Update patch set 1
Patch Set 1:

Good Morning,

I'm impasta coming from the XDA forums. I thought I should present a patchset about the "When I was trying to charge the phone while it is off, I get this confusing battery animation" item.

This is a set of 2 patches to enable lpm in Samsung Galaxy A7 2016.
The first one is all about prepping the SELinux not to deny lpm from running.
The second one is to include lpm in the proprietary-files.txt and set it up in the Android image.

Patch-set: 1
Reviewer: Gerrit User 26587 <26587@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>