a7xelte: Switch from passthrough to hwbinder GNSS HIDL
At random times, the GPS may crash on marshal_agps_ril_request_setid()
when using the "passthrough" transport.
The entie android system will also crash when this happens.

The reason is still unclear as to why it happens but switching over to
hwbinder seems to resolve the issue.

Please *NOTE* that for this switching to work,
`android.hardware.gnss@1.0-service` must also be built. This means that a
separate patch that adds this in the universal7580-common tree is needed.

If it is not built, the device may become stuck in the boot process, endlessly
waiting for the gnss HIDL service that will never ever be available.

Reference:
https://github.com/universal5433/android_device_samsung_gts2-common/commit/d37c8de84962a4dcb9bf3bd777a6242d61984c33

Signed-off-by: Mark Enriquez <bugmenot.oss@gmail.com>
Change-Id: I22366e1cb4eef815d9f852f4cdfbea425796b563